+++ openmetadata-ui/src/main/resources/ui/package.json
@@ -72,7 +72,7 @@
     "react-quill": "^1.3.5",
     "react-router-dom": "^5.2.0",
     "react-select": "^5.2.2",
-    "react-slack-chat": "^1.15.2",
+    "react-slack-chat": "https://github.com/vivekratnavel/react-slack-chat.git",
     "react-slick": "^0.28.1",
     "react-table": "^7.7.0",
     "react-tippy": "^1.4.0",
